requirements.txt
asgiref==3.10.0
attrs==25.4.0
cattrs==25.3.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
defusedxml==0.7.1
Django==5.2.7
django-ranged-response==0.2.0
django-simple-captcha==0.6.2
django-taggit==6.1.0
django-template-lsp==1.2.1
idna==3.11
jedi==0.19.2
lsprotocol==2023.0.1
oauthlib==3.3.1
parso==0.8.5
pillow==12.0.0
psycopg2==2.9.11
pycparser==2.23
pygls==1.3.1
PyJWT==2.10.1
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
social-auth-app-django==5.6.0
social-auth-core==4.8.1
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
